changed all unsigned longs to int, modified makefile for debug purposes, reallocate c matrix arrays in sp3mm_impl.f90

+++ base/serial/impl/sp3mm4amg/include/config.h
@@ -40,11 +40,11 @@ typedef enum{
 
 //
 typedef struct{
-	ushort gridRows;
-	ushort gridCols;
+	int gridRows;
+	int gridCols;
 	//TODO FULL CONFIG DOCCED HERE
 	ROW_MMSYM_IMPL_MODE symbMMRowImplID; //how to compute symb mul (if required)
-	uint threadNum;  //thread num to use in an OMP parallel region ...
+	int threadNum;  //thread num to use in an OMP parallel region ...
 	void* chunkDistrbFunc;  //CHUNKS_DISTR_INTERF func pntr
 } CONFIG;  
 ///Smart controls
